+++ dwds/web/reloader/require_restarter.dart
@@ -168,7 +168,7 @@ class RequireRestarter implements Restarter {
       );
 
   @override
-  Future<JSObject> hotReloadStart(String hotReloadSourcesPath) =>
+  Future<JSArray<JSObject>> hotReloadStart(String hotReloadSourcesPath) =>
       throw UnimplementedError(
         'Hot reload is not supported for the AMD module format.',
       );